feat: implement assignment dashboard logic and UI components for managing asset operations
