+++ src/components/features/operations/assignments/assignments-dashboard.tsx
@@ -1,61 +1,19 @@
 "use client";
 
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { toast } from "sonner";
 
-import {
-  sendAssignmentReminderAction,
-  requestAssetReturnAction,
-  markAssetReceivedAction
-} from "@/actions/assignments";
 import { AssignmentsPanels } from "./assignments-panels";
-import {
-  MultiAssetAssignmentModal,
-  type MultiAssetAssignmentItem,
-} from "./multi-asset-assignment-modal";
-import {
-  ProcessReturnModal,
-  type ReturnAssetItem,
-} from "./process-return-modal";
+import { MultiAssetAssignmentModal } from "./multi-asset-assignment-modal";
+import { ProcessReturnModal, type ReturnAssetItem } from "./process-return-modal";
 import { ModuleNavigationTabs } from "@/components/shared/module-navigation-tabs";
-import { Badge } from "@/components/ui/badge";
-import {
-  DataTable,
-  type DataTableSelectionAction,
-} from "@/components/shared/data-table";
-import { FilterBar, type AppliedFilter, type FilterFieldConfig } from "@/components/shared/filter-bar";
 import { TYPOGRAPHY_CLASSNAMES } from "@/components/shared/typography";
 import { TabsContent } from "@/components/ui/tabs";
-import type { ColumnDef } from "@tanstack/react-table";
-import type { RowSelectionState, OnChangeFn } from "@tanstack/react-table";
-import type { AssignmentsDashboardData, AssignmentsDashboardRow } from "@/lib/data/operations-assignments-repo";
+import type { AssignmentsDashboardData } from "@/lib/data/operations-assignments-repo";
 
-// --- Types ---
-
-type AssetAssignmentRow = {
-  assetId: string;
-  assetName: string;
-  serialNumber: string;
-  category: string;
-  status: string;
-  model: string;
-  brand: string;
-  owner: string;
-  group: string;
-  assignedTo: string;
-  department?: string;
-  assignedDate?: string;
-  expectedReturnDate?: string;
-  dateCreated: string;
-  updatedAt: string;
-  warranty: string;
-  note: string;
-  assetTag: string;
-  state: string;
-  assignmentId?: number;
-  lastRepaired?: string;
-};
+import { AssignmentsTable, type AssetAssignmentRow } from "./assignments-table";
+import { getAssignmentColumns } from "./assignments-columns";
+import { useAssignmentsDashboard } from "./use-assignments-dashboard";
 
 interface AssignmentsDashboardProps {
   data: AssignmentsDashboardData;
@@ -67,152 +25,43 @@ const tabs = [
   { id: "returned-assets", label: "Returned Assets" },
 ];
 
-// --- Helpers ---
-
-function formatDate(date: Date | string | null | undefined) {
-  if (!date) return "-";
-  const d = typeof date === "string" ? new Date(date) : date;
-  if (Number.isNaN(d.getTime())) return "-";
-  return d.toLocaleDateString("en-GB", {
-    day: "2-digit",
-    month: "short",
-    year: "numeric",
-  });
-}
-
-// --- Component ---
-
-interface AssignmentsTableProps {
-  rows: AssetAssignmentRow[];
-  columns: ColumnDef<AssetAssignmentRow>[];
-  selectionActions?: DataTableSelectionAction<AssetAssignmentRow>[];
-  showStatusColumn?: boolean;
-  searchValue: string;
-  onSearchChange: (value: string) => void;
-  filterFieldConfigs: FilterFieldConfig[];
-  appliedFilters: AppliedFilter[];
-  onApplyFilter: (filter: AppliedFilter) => void;
-  onClearFilter: (field: string) => void;
-  onClearAllFilters: () => void;
-  onRowClick: (row: AssetAssignmentRow, rowIndex: number) => void;
-  rowSelection: RowSelectionState;
-  onRowSelectionChange: OnChangeFn<RowSelectionState>;
-  disableSelectionHeader?: boolean;
-}
-
-function AssignmentsTable({
-  rows,
-  columns,
-  selectionActions,
-  showStatusColumn = false,
-  searchValue,
-  onSearchChange,
-  filterFieldConfigs,
-  appliedFilters,
-  onApplyFilter,
-  onClearFilter,
-  onClearAllFilters,
-  onRowClick,
-  rowSelection,
-  onRowSelectionChange,
-  disableSelectionHeader = false,
-}: AssignmentsTableProps) {
-  const tableColumns = showStatusColumn
-    ? columns
-    : columns.filter((col) => !("accessorKey" in col) || col.accessorKey !== "state");
-
-  return (
-    <div className="flex flex-col gap-4 flex-1 overflow-hidden min-h-0 mt-1">
-      <FilterBar
-        searchQuery={searchValue}
-        onSearchChange={onSearchChange}
-        searchPlaceholder="Search assets..."
-        fields={filterFieldConfigs}
-        appliedFilters={appliedFilters}
-        onApplyFilter={onApplyFilter}
-        onClearFilter={onClearFilter}
-        onClearAllFilters={onClearAllFilters}
-      />
-
-      <DataTable<AssetAssignmentRow, unknown>
-        columns={tableColumns}
-        data={rows}
-        onRowClick={onRowClick}
-        initialPageSize={10}
-        className="flex-1 min-h-0 rounded-lg border border-border"
-        selectionActions={selectionActions}
-        selectionLabel={(count) => `${count} Assets Selected`}
-        rowSelection={rowSelection}
-        onRowSelectionChange={onRowSelectionChange}
-        disableSelectionHeader={disableSelectionHeader}
-      />
-    </div>
-  );
-}
-
 export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
   const router = useRouter();
   const pathname = usePathname();
   const searchParams = useSearchParams();
-  const [isMultiAssignModalOpen, setIsMultiAssignModalOpen] = useState(false);
-  const [multiAssignAssets, setMultiAssignAssets] = useState<MultiAssetAssignmentItem[]>([]);
-  
+
   const [isProcessReturnModalOpen, setIsProcessReturnModalOpen] = useState(false);
   const [processReturnAsset, setProcessReturnAsset] = useState<ReturnAssetItem | null>(null);
-  
-  const [appliedFilters, setAppliedFilters] = useState<AppliedFilter[]>([]);
-  const [searchValue, setSearchValue] = useState("");
-  const [rowSelection, setRowSelection] = useState<RowSelectionState>({});
 
-  // 1. Panel State from URL (following the Registry Pattern)
+  const {
+    searchValue,
+    setSearchValue,
+    appliedFilters,
+    rowSelection,
+    setRowSelection,
+    applyFilter,
+    clearFilter,
+    clearAllFilters,
+    filterFieldConfigs,
+    assetRows,
+    filteredAvailableRows,
+    filteredAssignedRows,
+    filteredReturnedRows,
+    returnedRows,
+    selectionActionsAvailable,
+    selectionActionsAssigned,
+    isMultiAssignModalOpen,
+    setIsMultiAssignModalOpen,
+    multiAssignAssets,
+    setMultiAssignAssets,
+  } = useAssignmentsDashboard(data);
+
+  const columns = useMemo(() => getAssignmentColumns(), []);
+
+  // Panel State from URL (following the Registry Pattern)
   const activeAssetId = searchParams.get("id") || "";
   const currentPanel = searchParams.get("panel");
   const isPanelOpen = currentPanel === "record" && activeAssetId !== "";
-
-  // 2. Data Mapping
-  const mapRow = useCallback((asset: AssignmentsDashboardRow): AssetAssignmentRow => ({
-    assetId: asset.id,
-    assetName: asset.name ?? asset.assetTag,
-    serialNumber: asset.serialNumber ?? "-",
-    category: asset.category,
-    status: asset.status,
-    model: "-",
-    brand: "-",
-    owner: "-",
-    group: asset.pillar,
-    assignedTo: asset.assignedTo ?? "-",
-    department: asset.pillar,
-    assignedDate: formatDate(asset.assignedDate),
-    expectedReturnDate: formatDate(asset.expectedReturnDate),
-    dateCreated: formatDate(asset.createdAt),
-    updatedAt: formatDate(asset.updatedAt),
-    warranty: "-",
-    lastRepaired: "-",
-    note: asset.location ?? "-",
-    assetTag: asset.assetTag,
-    state: asset.state,
-    assignmentId: asset.assignmentId ?? undefined,
-  }), []);
-
-  const availableRows = useMemo<AssetAssignmentRow[]>(
-    () => data.available.map(mapRow),
-    [data.available, mapRow]
-  );
-
-  const assignedRows = useMemo<AssetAssignmentRow[]>(
-    () => data.assigned.map(mapRow),
-    [data.assigned, mapRow]
-  );
-
-  const returnedRows = useMemo<AssetAssignmentRow[]>(
-    () => data.returned.map(mapRow),
-    [data.returned, mapRow]
-  );
-
-  const assetRows = useMemo(
-    () => [...availableRows, ...assignedRows, ...returnedRows],
-    [availableRows, assignedRows, returnedRows]
-  );
 
   // Auto-open Process Return modal if navigating from Asset Details Panel
   useEffect(() => {
@@ -223,10 +72,8 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
         const row = returnedRows[targetRowIndex];
         
         const t = setTimeout(() => {
-          // 1. Select the row
           setRowSelection({ [targetRowIndex]: true });
           
-          // 2. Setup and open modal
           setProcessReturnAsset({
             assetId: row.assetId,
             assetTag: row.assetTag,
@@ -237,7 +84,6 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
           setIsProcessReturnModalOpen(true);
         }, 0);
 
-        // 3. Clean up the URL to prevent reopening on refresh
         const newParams = new URLSearchParams(searchParams.toString());
         newParams.delete("processReturnId");
         router.replace(`${pathname}?${newParams.toString()}`, { scroll: false });
@@ -245,138 +91,7 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
         return () => clearTimeout(t);
       }
     }
-  }, [searchParams, returnedRows, pathname, router]);
-
-  const categoryOptions = useMemo(() => {
-    const categories = new Set<string>();
-
-    for (const row of assetRows) {
-      if (row.category && row.category.trim().length > 0) {
-        categories.add(row.category);
-      }
-    }
-
-    return [...categories].sort((left, right) => left.localeCompare(right));
-  }, [assetRows]);
-
-  const statusOptions = useMemo(() => {
-    const statuses = new Set<string>();
-
-    for (const row of assetRows) {
-      if (row.state && row.state.trim().length > 0) {
-        statuses.add(row.state);
-      }
-    }
-
-    return [...statuses].sort((left, right) => left.localeCompare(right));
-  }, [assetRows]);
-
-  const filterFieldConfigs: FilterFieldConfig[] = useMemo(() => [
-    { value: 'Category', label: 'Category', options: categoryOptions },
-    { value: 'Status', label: 'Status', options: statusOptions },
-  ], [categoryOptions, statusOptions]);
-
-  const searchedAssetRows = useMemo(() => {
-    const query = searchValue.trim().toLowerCase();
-
-    if (!query) {
-      return assetRows;
-    }
-
-    return assetRows.filter((row) => {
-      return [
-        row.assetTag,
-        row.assetName,
-        row.serialNumber,
-        row.category,
-        row.status,
-        row.group,
-        row.assignedTo,
-        row.note,
-      ].some((field) => field.toLowerCase().includes(query));
-    });
-  }, [assetRows, searchValue]);
-
-  const filteredAssetRows = useMemo(() => {
-    if (appliedFilters.length === 0) {
-      return searchedAssetRows;
-    }
-
-    return searchedAssetRows.filter((row) => {
-      return appliedFilters.every((filter) => {
-        if (filter.field === 'Category') {
-          const matches = row.category === filter.value;
-          return filter.operator === "is" ? matches : !matches;
-        }
-        if (filter.field === 'Status') {
-          const matches = row.state === filter.value;
-          return filter.operator === "is" ? matches : !matches;
-        }
-        return true;
-      });
-    });
-  }, [searchedAssetRows, appliedFilters]);
-
-  const filteredAvailableRows = useMemo(
-    () => filteredAssetRows.filter((row) => availableRows.some((availableRow) => availableRow.assetId === row.assetId)),
-    [availableRows, filteredAssetRows]
-  );
-
-  const filteredAssignedRows = useMemo(
-    () => filteredAssetRows.filter((row) => assignedRows.some((assignedRow) => assignedRow.assetId === row.assetId)),
-    [assignedRows, filteredAssetRows]
-  );
-
-  const filteredReturnedRows = useMemo(
-    () => filteredAssetRows.filter((row) => returnedRows.some((returnedRow) => returnedRow.assetId === row.assetId)),
-    [filteredAssetRows, returnedRows]
-  );
-
-  const selectedAssignedRows = useMemo(() => {
-    return Object.keys(rowSelection)
-      .filter((key) => rowSelection[key])
-      .map((index) => filteredAssignedRows[Number(index)])
-      .filter(Boolean);
-  }, [rowSelection, filteredAssignedRows]);
-
-  const hasReminderCandidates = useMemo(() => {
-    return selectedAssignedRows.some(
-      (r) => ["pending approval", "overdue"].includes(r.state)
-    );
-  }, [selectedAssignedRows]);
-
-  const hasMarkReceivedCandidates = useMemo(() => {
-    return selectedAssignedRows.some(
-      (r) => ["requested", "overdue"].includes(r.state)
-    );
-  }, [selectedAssignedRows]);
-
-  const hasRequestAgainCandidates = useMemo(() => {
-    return selectedAssignedRows.some(
-      (r) => r.state === "requested"
-    );
-  }, [selectedAssignedRows]);
-
-  const hasReturnCandidates = useMemo(() => {
-    return selectedAssignedRows.some(
-      (r) => r.state === "assigned"
-    );
-  }, [selectedAssignedRows]);
-
-  const applyFilter = (nextFilter: AppliedFilter) => {
-    setAppliedFilters((currentFilters) => {
-      const withoutCurrentField = currentFilters.filter((f) => f.field !== nextFilter.field);
-      return [...withoutCurrentField, nextFilter];
-    });
-  };
-
-  const clearFilter = (field: string) => {
-    setAppliedFilters((currentFilters) => currentFilters.filter((f) => f.field !== field));
-  };
-
-  const clearAllFilters = () => {
-    setAppliedFilters([]);
-  };
+  }, [searchParams, returnedRows, pathname, router, setRowSelection]);
 
   const handleClosePanel = () => {
     const params = new URLSearchParams(searchParams.toString());
@@ -391,196 +106,17 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
     [assetRows, activeAssetId]
   );
 
-  const selectionActionsAvailable = useMemo<DataTableSelectionAction<AssetAssignmentRow>[]>(
-    () => [
-      {
-        id: "assign-assets",
-        label: "Assign Assets",
-        tone: "primary",
-        onClick: (selectedRows) => {
-          setMultiAssignAssets(
-            selectedRows.map((row) => ({
-              assetId: row.assetId,
-              assetTag: row.assetTag,
-              assetName: row.assetName,
-              assetGroup: row.group,
-            }))
-          );
-          setIsMultiAssignModalOpen(true);
-        },
-      },
-    ],
-    []
-  );
-
-  const selectionActionsAssigned = useMemo<DataTableSelectionAction<AssetAssignmentRow>[]>(
-    () => {
-      const actions: DataTableSelectionAction<AssetAssignmentRow>[] = [];
-
-      if (hasMarkReceivedCandidates) {
-        actions.push({
-          id: "mark-received",
-          label: "Returned",
-          tone: "secondary",
-          onClick: async (selectedRows) => {
-            const ids = selectedRows
-              .filter((r) => ["requested", "overdue"].includes(r.state))
-              .map((r) => r.assignmentId)
-              .filter((id): id is number => id !== undefined);
-
-            if (ids.length === 0) return;
-
-            const result = await markAssetReceivedAction(ids);
-            if (result.success) {
-              toast.success(`${ids.length} assets marked as returned`);
-              setRowSelection({});
-            } else {
-              toast.error(result.error || "Failed to mark as returned");
-            }
-          },
-        });
-      }
-
-      if (hasRequestAgainCandidates) {
-        actions.push({
-          id: "request-again",
-          label: "Request Again",
-          tone: "secondary",
-          onClick: async (selectedRows) => {
-            const ids = selectedRows
-              .filter((r) => r.state === "requested")
-              .map((r) => r.assignmentId)
-              .filter((id): id is number => id !== undefined);
-
-            if (ids.length === 0) return;
-
-            const result = await requestAssetReturnAction(ids);
-            if (result.success) {
-              toast.success(`Return re-requested for ${ids.length} assets`);
-              setRowSelection({});
-            } else {
-              toast.error(result.error || "Failed to re-request return");
-            }
-          },
-        });
-      }
-
-      if (hasReminderCandidates) {
-        actions.push({
-          id: "send-reminder",
-          label: "Send Reminder",
-          tone: "primary",
-          onClick: async (selectedRows) => {
-            const reminderIds = selectedRows
-              .filter((r) => ["pending approval", "overdue"].includes(r.state))
-              .map((r) => r.assignmentId)
-              .filter((id): id is number => id !== undefined);
-
-            if (reminderIds.length === 0) return;
-
-            const result = await sendAssignmentReminderAction(reminderIds);
-            if (result.success) {
-              toast.success(`Reminder sent for ${reminderIds.length} assets`);
-              setRowSelection({});
-            } else {
-              toast.error(result.error || "Failed to send reminders");
-            }
-          },
-        });
-      }
-
-      if (hasReturnCandidates) {
-        actions.push({
-          id: "request-return",
-          label: "Request Return",
-          tone: "primary",
-          onClick: async (selectedRows) => {
-            const returnIds = selectedRows
-              .filter((r) => r.state === "assigned")
-              .map((r) => r.assignmentId)
-              .filter((id): id is number => id !== undefined);
-
-            if (returnIds.length === 0) return;
-
-            const result = await requestAssetReturnAction(returnIds);
-            if (result.success) {
-              toast.success(`Return requested for ${returnIds.length} assets`);
-              setRowSelection({});
-            } else {
-              toast.error(result.error || "Failed to request return");
-            }
-          },
-        });
-      }
-
-      return actions;
-    },
-    [hasReminderCandidates, hasMarkReceivedCandidates, hasRequestAgainCandidates, hasReturnCandidates]
-  );
-
-  // 3. Column Definitions for the Hardware Registry View
-  const columns = useMemo<ColumnDef<AssetAssignmentRow>[]>(() => [
-    {
-      accessorKey: "assetId",
-      header: "Asset ID",
-      cell: ({ row }) => <span className="font-medium text-foreground">{row.original.assetTag}</span>,
-    },
-    {
-      accessorKey: "assetName",
-      header: "Asset Name",
-    },
-    {
-      accessorKey: "serialNumber",
-      header: "Serial Number",
-      cell: ({ row }) => <span className="text-muted-foreground font-mono">{row.original.serialNumber}</span>,
-    },
-    {
-      accessorKey: "category",
-      header: "Category",
-      cell: ({ row }) => (
-        <Badge variant="outline" className="h-5 rounded-full border-border bg-muted px-2 text-[11px] font-medium text-muted-foreground">
-          {row.original.category}
-        </Badge>
-      ),
-    },
-    {
-      accessorKey: "state",
-      header: "Status",
-      cell: ({ row }) => {
-        const state = row.original.state;
-        const colors: Record<string, string> = {
-          'pending approval': "bg-amber-50 text-amber-700 border-amber-200",
-          'assigned': "bg-emerald-50 text-emerald-700 border-emerald-200",
-          'overdue': "bg-rose-50 text-rose-700 border-rose-200",
-          'requested': "bg-blue-50 text-blue-700 border-blue-200",
-          'returned': "bg-muted text-foreground border-border",
-        };
-        const colorClass = colors[state] || "bg-muted text-foreground border-border";
-
-        return (
-          <Badge variant="outline" className={`h-5 rounded-full px-2 text-[11px] font-medium capitalize ${colorClass}`}>
-            {state}
-          </Badge>
-        );
-      },
-    },
-  ], []);
-
-  // 4. Handlers to trigger the Slide Panel via URL
   const handleRowClick = (row: AssetAssignmentRow) => {
     const params = new URLSearchParams(searchParams.toString());
     params.set("panel", "record");
     params.set("id", row.assetTag);
-    // Mimic the animation toggle logic from RegistryClient
     params.set("animate", isPanelOpen ? "0" : "1");
     router.push(`${pathname}?${params.toString()}`, { scroll: false });
   };
 
   const handleReturnedAssetClick = (row: AssetAssignmentRow, rowIndex: number) => {
-    // 1. Check the checkbox of the clicked row exclusively
     setRowSelection({ [rowIndex]: true });
     
-    // 2. Open the process return modal
     setProcessReturnAsset({
       assetId: row.assetId,
       assetTag: row.assetTag,
@@ -593,7 +129,6 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
 
   const handleMultiAssignModalOpenChange = (open: boolean) => {
     setIsMultiAssignModalOpen(open);
-
     if (!open) {
       setMultiAssignAssets([]);
     }
@@ -609,7 +144,7 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
             </h1>
           </div>
 
-            <div className="flex flex-col min-h-0 flex-1 overflow-hidden">
+          <div className="flex flex-col min-h-0 flex-1 overflow-hidden">
             <ModuleNavigationTabs
               tabs={tabs}
               defaultTab="available-assets"
@@ -694,7 +229,7 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
           setIsProcessReturnModalOpen(open);
           if (!open) {
             setProcessReturnAsset(null);
-            setRowSelection({}); // Clear selection when modal closes
+            setRowSelection({});
           }
         }}
       />
